refactored new plan design. changed exercise selection method
