+++ app/main.js
@@ -16,7 +16,7 @@ import { fetchCurrentPatient } from './reducers/currentpatient';
 import Home from './components/home/Home';
 import App from './components/App';
 import AddPatientContainer from './components/patients/AddPatientContainer';
-import newPlansContainer from './components/plans/newPlanContainer';
+import NewPlanContainer from './components/plans/NewPlanContainer';
 import PlanConfirmContainer from './components/plans/PlanConfirmContainer';
 import Plan from './components/plan/PatientPlan';
 import Counter from './components/plan/Counter';
@@ -89,7 +89,7 @@ render (
         <Route path="/dashboard" component={ Dashboard } onEnter={ patientPlanEnter } />
         <Route path="/patients" component={ PatientListContainer } onEnter={ patientsListEnter } />
         <Route path="/patients/new" component={ AddPatientContainer } />
-        <Route path="/patients/:patientId/plans/new" component={newPlansContainer} onEnter={newPlanEnter} />
+        <Route path="/patients/:patientId/plans/new" component={NewPlanContainer} onEnter={newPlanEnter} />
         <Route path="/patients/dashboard" component={ Dashboard } />
         <Route path="/messages" component={ ChatRoom } />
         <Route path="/patients/:patientId/plans/current" component={ CurrentPlan } onEnter={therapistPlanEnter} />
