+++ src/stylesheets/info-panel.scss
@@ -1,0 +1,40 @@
+.info-panel { // Paper container
+  min-width: 25% !important;
+  max-width: 35% !important;
+  margin-left: 20px;
+  .info-content {
+    display: flex;
+    flex-direction: column;
+    padding: 10px;
+    > img {
+      width: auto;
+      height: auto;
+      display: flex;
+      border-radius: 50%;
+      align-self: center;
+    }
+    > img + .info-item {
+      padding-top: 10px;
+    }
+    > .info-item + .info-item {
+      padding: 5px 0;
+    }
+  }
+}
+
+@media (max-width: $lg) {
+  .info-panel { // Paper container
+    max-width: initial !important;
+    margin-left: 0;
+    .info-content {
+      display: flex;
+      flex-direction: row;
+      flex-wrap: wrap;
+      justify-content: space-around;
+      align-items: center;
+      > .info-item + .info-item, > img + .info-item {
+        padding: 0;
+      }
+    }
+  }
+}
