--- src/stylesheets/newplan.scss
@@ -1,11 +1,0 @@
-@import "colors";
-
-#newPlan .plan-options {
-  display: flex;
-}
-
-img .profile {
-  height:100px;
-  width: 100px;
-
-}
