+++ src/stylesheets/style.scss
@@ -6,11 +6,10 @@
 @import "video";
 @import "counter";
 @import "chat";
-@import "newplan";
+@import "new-plan";
 @import "patientlist";
 @import "info-item";
-
->>>>>>> master
+@import "info-panel";
 
 body,html {
   width : 100%;
